CNN foundation
Made a start with the CNN foundation. Was not able to test the current performance, as the training would not load, which I still don't know why. The layers should currently work, but are not likely to be effective. Furthermore, I did not see any problems with the crossentropy loss function